Add ignore option for checks

+++ scripts/exp-deploy-set-tag.sh
@@ -1,4 +1,7 @@
 #!/bin/sh -e
 
-git tag -f deployed
-git push --force origin deployed
+# Skip tagging if we are not deploying to all servers.
+if [[ -z $EXP_DEPLOY_SERVERS ]]; then
+  git tag -f deployed
+  git push --force origin deployed
+fi
